tidying up login code.
this is partly because i realized it was really gnarly when i looked at
it this afternoon while trying to add a really gross form. It was so
complicated that even i was unhappy trying to make any change.

But also i think the previous version was making lots of stuff really
hard to follow, so i thought it would help to rearrange this so that
you can make changes to App or AppBody more easily.

This is still a bunch of code, but i think it is more clear-headed
because i am writing it in the light of day and not at like 1am.