Abstracts data fetching into a react component
This seems crazy (and it sort of is) but part of the reason i'm doing
it like this is to also point out that react is really about creating
a visual hierarchy to your code in a way that aids abstraction.

One way of thinking about firebase is like an <input />. It has all the
data, so when we ask it for data, we can pretend like it is just
updating a form component and updating our app's central redux store.
Just like a form <input />, it doesn't need to actually feed the data to
its children or anything because it is happening asynchronously.

As long as the `<FirebaseData />` component remains mounted, it will
call its update handlers whenever firebase sees new data or it changes